mitmproxy_addon/pyproject.toml
[project]
name = "proxylens-mitmproxy"
version = "0.1.0"
description = "mitmproxy addon client for ProxyLens Server"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mitmproxy==12.2.1",
    "python-ulid>=3.1.0",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
    "black>=25.1,<26",
    "pytest==9.0.2",
    "pytest-watch>=4.2,<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/proxylens_mitmproxy"]

[tool.pytest.ini_options]
testpaths = ["src", "tests"]
addopts = "-q"
python_files = ["*_test.py", "test_*_integration.py"]
filterwarnings = [
    "ignore:tagMap is deprecated\\. Please use TAG_MAP instead\\.:DeprecationWarning",
    "ignore:typeMap is deprecated\\. Please use TYPE_MAP instead\\.:DeprecationWarning",
]
